Refactor logging and add models and config modules
