+++ tests/test_logging.py
@@ -4,13 +4,13 @@ import argparse
 import logging
 from unittest.mock import MagicMock, patch
 
-from cf_ips_to_hcloud_fw.logging import init_logging
+from cf_ips_to_hcloud_fw.logging import setup_logging
 
 
 @patch("logging.basicConfig")
-def test_init_logging_debug(mock_basic_config: MagicMock) -> None:
+def test_setup_logging_debug(mock_basic_config: MagicMock) -> None:
     args = argparse.Namespace(debug=True)
-    init_logging(args)
+    setup_logging(args)
     mock_basic_config.assert_called_once_with(
         level=logging.getLevelName(logging.DEBUG),
         format="%(asctime)s %(levelname)-8s "
@@ -19,9 +19,9 @@ def test_init_logging_debug(mock_basic_config: MagicMock) -> None:
 
 
 @patch("logging.basicConfig")
-def test_init_logging_info(mock_basic_config: MagicMock) -> None:
+def test_setup_logging_info(mock_basic_config: MagicMock) -> None:
     args = argparse.Namespace(debug=False)
-    init_logging(args)
+    setup_logging(args)
     mock_basic_config.assert_called_once_with(
         level=logging.getLevelName(logging.INFO),
         format="%(asctime)s %(levelname)-8s %(message)s",
